Simplified guard clause

+++ UltimateMahjongConnect.Application.Net/Services/GamerService.cs
@@ -1,11 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using UltimateMahjongConnect.Application.Interface;
+﻿using UltimateMahjongConnect.Application.Interface;
 using UltimateMahjongConnect.Application.DTO;
 using UltimateMahjongConnect.Domain.Models;
 using AutoMapper;
-using System.Data.Common;
 
 namespace UltimateMahjongConnect.Application.Services
 {
@@ -55,18 +51,14 @@ namespace UltimateMahjongConnect.Application.Services
 
         public async Task<GamerDTO?> GetGamerByIdAsync(int id)
         {
-            if (id <= 0) throw new ArgumentException("L'identifiant du joueur doit être supérieur à 0", nameof(id));
-            else
+            try
             {
-                try
-                {
-                    var gamer = await _gamerRepository.GetGamerByIdAsync(id);
-                    return gamer != null ? _mapper.Map<GamerDTO>(gamer) : null;
-                }
-                catch (Exception ex)
-                {
-                    throw new ApplicationException($"Une erreur est survenue lors de la récupération du joueur avec le pseudonyme '{id}'", ex);
-                }
+                var gamer = await _gamerRepository.GetGamerByIdAsync(id);
+                return gamer != null ? _mapper.Map<GamerDTO>(gamer) : null;
+            }
+            catch (Exception ex)
+            {
+                throw new ApplicationException($"Une erreur est survenue lors de la récupération du joueur avec le pseudonyme '{id}'", ex);
             }
         }
 
@@ -99,13 +91,9 @@ namespace UltimateMahjongConnect.Application.Services
             {
                 throw new ArgumentNullException(nameof(gamerDTO));
             }
-            if (gamerDTO.Id <= 0)
-            {
-                throw new ArgumentException("L'identifiant du joueur doit être supérieur à 0", nameof(gamerDTO));
-            }
             try
             {
-                var foundGamer = await GetGamerByIdAsync(gamerDTO.Id);
+                var foundGamer = await _gamerRepository.GetGamerByIdAsync(gamerDTO.Id);
                 if (foundGamer != null)
                 {
                     var gamer = _mapper.Map<Gamer>(gamerDTO);
@@ -124,18 +112,22 @@ namespace UltimateMahjongConnect.Application.Services
 
         public async Task DeleteGamerAsync(GamerDTO gamerDTO)
         {
+            if (gamerDTO == null)
+            {
+                throw new ArgumentNullException(nameof(gamerDTO));
+            }
             try
             {
-                var foundGamer = await GetGamerByIdAsync(gamerDTO.Id);
+                var foundGamer = await _gamerRepository.GetGamerByIdAsync(gamerDTO.Id);
                 if (foundGamer != null)
                 {
                     var gamer = _mapper.Map<Gamer>(gamerDTO);
                     await _gamerRepository.DeleteGamerAsync(gamer);
                 }
-            } 
-            catch(Exception ex)
+            }
+            catch (Exception ex)
             {
-                throw new ApplicationException();
+                throw new ApplicationException("Une erreur est survenue lors de la suppression du joueur", ex);
             }
         }
